// Sale.Check.In.WebApp/ClientApp/src/app/model/check-in-save.model.ts
export interface CheckInSaveModel {
  userId: number;
  latitude: number;
  longitude: number;
  shopName: string;
  shopType: string;
  customerType: string;
  province: string;
  amphur: string;
  district: string;
  reportType: string;
  note: string;
  receiptFile: string;
  mimeType: string;
}